requirements.txt
langchain_openai 
langchain_core
python-dotenv
langchain_community
langserve
Flask
diffusers
torch
transformers